Silence warning

+++ cpp/src/lib/Utils.cpp
@@ -14,7 +14,7 @@
 // Don't be weird: return an int, not size_t
 //
 int Utils::parseInteger(std::string s, size_t base) {
-    return std::stoi(s, nullptr, base);
+    return std::stoi(s, nullptr, static_cast<int>(base));
 }
